Move new_module to separate component to fix issue during build.
